デルタタイム等の統一、スカイドームの作成、実装

+++ ShootingGame_0519/IScene.h
@@ -15,10 +15,10 @@ public:
 
 	//-------------一連の流れに関しての純粋仮想関数-------------
 	//Update関数(uint64_t deltaは前回からの経過時間)
-	virtual void Update(uint64_t delta) = 0; 
+	virtual void Update(float delta) = 0; 
 
 	//Draw関数(uint64_t deltaは前回からの経過時間)
-	virtual void Draw(uint64_t delta) = 0;	 
+	virtual void Draw(float delta) = 0;	 
 
 	//Init関数
 	virtual void Init() = 0;
